hotvpn/requirements.txt
Django==3.0.6
ansible==2.9.9
ansible-runner==1.4.6
ansible-ssh==0.9.1
cryptography==2.7
celery==4.4.7
channels==2.4.0
channels-redis==2.4.2
django-session-timeout==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
django-crispy-forms==1.9.2
gunicorn==20.0.4
paramiko==2.7.1
redis==3.5.3
scp==0.13.2
websocket-client==0.57.0
psutil==5.7.0
py-cpuinfo==7.0.0
